+++ frontend/src/components/Timer.js
@@ -2,8 +2,23 @@ import React from "react";
 import { useState, useEffect } from "react";
 import styles from "./Timer.module.css";
 
-const Timer = function ({ sec, background }) {
-  const record = [];
+const Timer = function () {
+  const [sec, setSec] = useState(10);
+  const [background, setbackground] = useState("blue");
+
+  useEffect(() => {
+    if (sec <= 5) {
+      setbackground("red");
+    }
+    const timer = setTimeout(() => {
+      setSec(sec - 1);
+    }, 1000);
+
+    if (sec === 0) {
+      clearTimeout(timer);
+      //setSec("You finish!!");
+    }
+  });
 
   return (
     <div className={`${styles.timer} ${styles[`${background}`]}`}>
